@@ -1,22 +1,25 @@
 {
     "name": "lazypdo/lazypdo",
     "description": "LazyPDO is a set of wrappers over PHP's standard PDO and PDOStatement classes. It enables lazy loading, serialization and decoration.",
-    "keywords": ["test", "unit test", "pdo", "mock", "lazyload", "lazy load", "serializable", "serialize"],
+    "keywords": [
+        "test",
+        "unit test",
+        "pdo",
+        "mock",
+        "lazyload",
+        "lazy load",
+        "serializable",
+        "serialize"
+    ],
     "homepage": "https://github.com/lazypdo/lazypdo",
     "license": "MIT",
-    "authors": [
-        {
-            "name": "Alexey Karapetov",
-            "email": "karapetov@gmail.com",
-            "role": "Developer"
-        }
-    ],
     "require": {
-        "php": ">=5.3.0 <7.4"
+        "php": ">=7.4",
+        "ext-pdo": "*"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8 || ^5.0",
-        "squizlabs/php_codesniffer": "^2.0 || ^3.0"
+        "phpunit/phpunit": "^6.0",
+        "squizlabs/php_codesniffer": "^3.0"
     },
     "autoload": {
         "psr-4": {
