@@ -1,14 +1,5 @@
 <?php
 
-/**
- * This file is part of LazyPDO.
- *
- * (c) Alexey Karapetov <karapetov@gmail.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
 namespace LazyPDO;
 
 use PDOStatement;
@@ -18,7 +9,7 @@ class SimplePDOStatementDecorator extends PDOStatementDecorator
     /**
      * @var PDOStatement
      */
-    private $pdoStatement;
+    private PDOStatement $pdoStatement;
 
     /**
      * SimplePDOStatementDecorator constructor.
