@@ -1,19 +1,11 @@
 <?php
 
-/**
- * This file is part of LazyPDO.
- *
- * (c) Alexey Karapetov <karapetov@gmail.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
 namespace LazyPDO\Test;
 
 use LazyPDO\LazyPDO;
+use PHPUnit\Framework\TestCase;
 
-class LazyPDOTest extends \PHPUnit_Framework_TestCase
+class LazyPDOTest extends TestCase
 {
     private $pdo;
 
